backend/foodgram_backend/requirements.txt
asgiref==3.4.1
coverage==6.0.2
Django==3.2.8
django-colorfield==0.4.5
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
Pillow==8.4.0
psycopg2-binary==2.9.1
PyJWT==2.3.0
python-dotenv==0.19.1
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
